+++ examples/matrixmultiplication4/src/at/illecker/rootbeer/examples/matrixmultiplication4/MatrixMultiplication4Kernel.java
@@ -259,7 +259,7 @@ public class MatrixMultiplication4Kernel implements Kernel {
     Context context = rootbeer.createDefaultContext();
     Stopwatch watch = new Stopwatch();
     watch.start();
-    rootbeer.run(kernel, new ThreadConfig(blockSize, gridSize, blockSize
+    rootbeer.run(kernel, new ThreadConfig(blockSize, gridSize, (long) blockSize
         * gridSize), context);
     watch.stop();
