+++ examples/piestimator/src/at/illecker/rootbeer/examples/piestimator/PiEstimatorKernel.java
@@ -129,7 +129,7 @@ public class PiEstimatorKernel implements Kernel {
     Context context = rootbeer.createDefaultContext();
     Stopwatch watch = new Stopwatch();
     watch.start();
-    rootbeer.run(kernel, new ThreadConfig(blockSize, gridSize, blockSize
+    rootbeer.run(kernel, new ThreadConfig(blockSize, gridSize, (long) blockSize
         * gridSize), context);
     watch.stop();
